requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pillow>=9.0.0
scikit-learn>=1.0.0
transformers>=4.20.0
timm>=0.6.0
tensorboard>=2.9.0
pandas>=1.3.0
tqdm>=4.62.0

# Inference & Serving
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
gradio>=4.0.0
